requirements.txt
requests
assertlib
pytest
allure-pytest
pytest-xdist
pytest-variables
pytest-rerunfailures
delayed-assert
